Added CPK Tests and Refactoring
