+++ sdk/storage/azblob/zc_retry_reader.go
@@ -56,6 +56,9 @@ type RetryReaderOptions struct {
 	// from the same "thread" (goroutine) as Read.  Concurrent Close calls from other goroutines may instead produce network errors
 	// which will be retried.
 	TreatEarlyCloseAsError bool
+
+	CpkInfo      *CpkInfo
+	CpkScopeInfo *CpkScopeInfo
 }
 
 // retryReader implements io.ReaderCloser methods.
